Yael's R code README.md
This code is for a question within my lab that is due tomorrow and I am still beyond screwed :(

PROMPT:
Exploratory Data Analysis The file fev.dat contains data related to pulmonary function and smoking
status for a sample of youth. The file fev-desc.txt gives a description of the data. Use the descriptive
and graphical techniques discussed in class to analyize these data. Prepare a short write-up discussing
your conclusions.